Moved module runner to separate file and moved entry path into env
