+++ examples/worker/worker-a/wrangler.toml
@@ -1,3 +1,6 @@
 compatibility_date = "2024-09-09"
 
-services = [{ binding = "WORKER_B", service = "worker_b" }]
+services = [
+	{ binding = "WORKER_B", service = "worker_b" },
+	# { binding = "TEST", service = "worker_b", entrypoint = "Test" },
+]
