+++ packages/vite-plugin-cloudflare/src/runner/env.ts
@@ -1,0 +1,10 @@
+export interface WrapperEnv {
+	__VITE_ROOT__: string;
+	__VITE_ENTRY_PATH__: string;
+	__VITE_FETCH_MODULE__: {
+		fetch: (request: Request) => Promise<Response>;
+	};
+	__VITE_UNSAFE_EVAL__: {
+		eval: (code: string, filename: string) => Function;
+	};
+}
